[Comments] useQuillContent hook
